gRPC: fix replacement char encoding (#55327)
The byte sequence I used to replace invalid byte sequences was itself an
invalid byte sequence 🤦 . `0xFFFD` is the code point for the unicode
replacement character, not the byte sequence in UTF8. Go supports utf8
characters in strings, so to avoid the error, I just pasted the raw
character itself.