sveltekit: Group routes that require a valid revision (#55012)
This PR makes use of SvelteKit's error handling for repository related
errors.

I added a
[matcher](https://kit.svelte.dev/docs/advanced-routing#matching) for
`repo` to ensure that paths to non-existing pages (such as
`/github.com/sourcegraph/sourcegraph/-/doesntexist`) are handled
correctly. Without this matcher, SvelteKit would extract the whole path
as `repo` name and we would report it as "repo not found error".

I added the new
[group](https://kit.svelte.dev/docs/advanced-routing#advanced-layouts-group),
`(validrev)` and moved all existing pages underneath it. The layout
loader in this group verifies that the requested revision is valid. If
it isn't it throws an error and subsequently renders the closest error
page. If it is it makes the (valid!) revision information available to
the contained pages. That means all these pages can now assume that
`resolvedRevision` is a valid object and we can remove a bunch of `null`
and `isErrorLike` checks from those files.

NOTE: Because I moved and changed the loader files at the same time, Git
reports them as deleted and new files unfortunately.

In a future PR we can add the repo settings pages, which, according to
our existing app, should also be accessible when the requested revision
is not valid. Those paths would be created next to (not inside)
`(validrev)`.



![2023-07-17_14-39](https://github.com/sourcegraph/sourcegraph/assets/179026/66d1270f-ea44-418c-9edf-ede3b2da5128)

![2023-07-17_14-40](https://github.com/sourcegraph/sourcegraph/assets/179026/f5cfd053-5d55-459d-91b8-2c7e68bcfd16)


## Test plan

- Navigate to URL with invalid repo name
- Navigate to URL with invalid revision
- Navigate to invalid repo path
- Navigate to valid repo paths