sveltekit: Add line selection to blob view (#55037)
This PR adds line selection to the blob view and to search results. The
search results implementation is a bit hacky: I'm listening for click
events on line numbers and programmatically trigger a page navigation.
Eventually we need to find a better approach (I also didn't want to do
the same as we do in the current app, since this interferes a bit with
text selection).

NOTE: Since the whole blob view page is now scrollable, the logic that
we currently use to determine whether or not to scroll a selected line
into view doesn't work. That logic depends on CodeMirror's scroll DOM to
be scrollable, which is not the case here. This should be improved in a
separate PR.



https://github.com/sourcegraph/sourcegraph/assets/179026/96baae33-de8c-4254-a9ef-341314c3fead



## Test plan

- Click line number in search results -> opens blob view at the clicked
line
- Click line number in blob view -> mark line as selected and update URL